Add files via upload
This program takes in to numbers that represent the endpoints of an interval. This interval is {a, b}. It confirms that 10 <= a <=b < 10000. It then determines if each number in the interval is a "mountain" or a "valley". A mountain is defined as a number such that, in the 2 digit case, digit1 < digit2 and a valley is defined as digit1 > digit2. A number can be neither a mountain nor a valley. The program makes use of multiple functions to classify each number in the interval and keeps track of how many mountains or valleys there are. The output is then displayed. This program works by number slicing and comparing digits. It works on up to four digit numbers as the endpoint is a maximum of 9999.